Fixes #17323 - Allow SSH provisioning with CR cert
Problem:

SSH provisioning does not work. Only PXE provisioning is possible
through smart proxies in an Azure network.
The reason is that the setSSHWaitForResponse does not come with support
for using keys other than the foreman user ones.

Solution:

The certificate passed on to the Azure compute resource can be used to
SSH provision these machines as soon as they get a public IP.
The vm_user attribute that's passed on to Azure via fog will be the
username, and sudo is allowed in most Azure images.